Create googlemap.php
The idea is simple. Make a request to Google Maps server, and it will return an XML (or JSON). Then, parse the XML to get the latitude and longitude.

Now we have latitude and longitude of the given address.Our next step is to display the Google Map.This can be done using  Java Script.Code.